Initial refactoring to separate OC
