+++ monitoring/monitoring.go
@@ -15,20 +15,12 @@
 package monitoring
 
 import (
-	"context"
-	"fmt"
-	"math"
 	"sync"
-
-	"go.opencensus.io/metric"
-	"go.opencensus.io/metric/metricdata"
-	"go.opencensus.io/metric/metricproducer"
-	"go.opencensus.io/stats"
-	"go.opencensus.io/stats/view"
-	"go.opencensus.io/tag"
-
-	"istio.io/pkg/log"
 )
+
+// OpenCensus independent metrics
+// Removed unused code:
+// - Metric.Decrement
 
 type (
 	// A Metric collects numerical observations.
@@ -43,25 +35,42 @@ type (
 		// this is equivalent to subtracting -1 to the current value. For Gauges,
 		// this is equivalent to setting the value to -1. For Distributions,
 		// this is equivalent to making an observation of value -1.
-		Decrement()
+		//
+		// Not used in istio, removed pending new interface.
+		//Decrement()
 
 		// Name returns the name value of a Metric.
+		// TODO: internal use only
 		Name() string
 
 		// Record makes an observation of the provided value for the given measure.
+		// Majority of Istio is setting this to an int
 		Record(value float64)
 
 		// RecordInt makes an observation of the provided value for the measure.
-		RecordInt(value int64)
+		// Not actually used in Istio.
+		//RecordInt(value int64)
+
+		// Prometheus uses ConterVec and With(map[string]string) returning Counter
+		//
+		// Expvar doesn't support labels - but can be used in the name, creating a new expvar in a map
+		// The labels are really made part of the name when exporting and is the name of the TS
+		//
+		// Otel uses attribute.Int("name", val) and similar - from the stable package -
+		// but doesn't create a new Metric, it is an option to Add().
+		//
+		// Istio only uses string/string - so using the pattern from slog works.
+		// Once we adopt slog, we can start using slog.Attr pattern
 
 		// With creates a new Metric, with the LabelValues provided. This allows creating
 		// a set of pre-dimensioned data for recording purposes. This is primarily used
 		// for documentation and convenience. Metrics created with this method do not need
 		// to be registered (they share the registration of their parent Metric).
-		With(labelValues ...LabelValue) Metric
+		With(labelValues ...string) Metric
 
 		// Register configures the Metric for export. It MUST be called before collection
 		// of values for the Metric. An error will be returned if registration fails.
+		// TODO: internal use only
 		Register() error
 	}
 
@@ -96,15 +105,11 @@ type (
 	DerivedOptions func(*derivedOptions)
 
 	// A Label provides a named dimension for a Metric.
-	Label tag.Key
-
-	// A LabelValue represents a Label with a specific value. It is used to record
-	// values for a Metric.
-	LabelValue tag.Mutator
+	//Label Key
 
 	options struct {
 		unit     Unit
-		labels   []Label
+		labels   []string // Label
 		useInt64 bool
 	}
 
@@ -112,13 +117,10 @@ type (
 		labelKeys []string
 		valueFn   func() float64
 	}
-
-	// RecordHook has a callback function which a measure is recorded.
-	RecordHook interface {
-		OnRecordFloat64Measure(f *stats.Float64Measure, tags []tag.Mutator, value float64)
-		OnRecordInt64Measure(i *stats.Int64Measure, tags []tag.Mutator, value int64)
-	}
 )
+
+func (dm *disabledMetric) ValueFrom(valueFn func() float64, labelValues ...string) {
+}
 
 // Decrement implements Metric
 func (dm *disabledMetric) Decrement() {}
@@ -143,35 +145,19 @@ func (dm *disabledMetric) Register() error {
 }
 
 // With implements Metric
-func (dm *disabledMetric) With(labelValues ...LabelValue) Metric {
+func (dm *disabledMetric) With(labelValues ...string) Metric {
 	return dm
 }
 
 var _ Metric = &disabledMetric{}
 
 var (
-	recordHooks     map[string]RecordHook
 	recordHookMutex sync.RWMutex
-
-	derivedRegistry = metric.NewRegistry()
 )
-
-func init() {
-	recordHooks = make(map[string]RecordHook)
-	// ensures exporters can see any derived metrics
-	metricproducer.GlobalManager().AddProducer(derivedRegistry)
-}
-
-// RegisterRecordHook adds a RecordHook for a given measure.
-func RegisterRecordHook(name string, h RecordHook) {
-	recordHookMutex.Lock()
-	defer recordHookMutex.Unlock()
-	recordHooks[name] = h
-}
 
 // WithLabels provides configuration options for a new Metric, providing the expected
 // dimensions for data collection for that Metric.
-func WithLabels(labels ...Label) Options {
+func WithLabels(labels ...string) Options {
 	return func(opts *options) {
 		opts.labels = labels
 	}
@@ -212,20 +198,20 @@ func WithValueFrom(valueFn func() float64) DerivedOptions {
 	}
 }
 
-// Value creates a new LabelValue for the Label.
-func (l Label) Value(value string) LabelValue {
-	return tag.Upsert(tag.Key(l), value)
-}
-
-// MustCreateLabel will attempt to create a new Label. If
-// creation fails, then this method will panic.
-func MustCreateLabel(key string) Label {
-	k, err := tag.NewKey(key)
-	if err != nil {
-		panic(fmt.Errorf("could not create label %q: %v", key, err))
-	}
-	return Label(k)
-}
+//// Value creates a new LabelValue for the Label.
+//func (l Label) Value(value string) LabelValue {
+//	return tag.Upsert(tag.Key(l), value)
+//}
+//
+//// MustCreateLabel will attempt to create a new Label. If
+//// creation fails, then this method will panic.
+//func MustCreateLabel(key string) Label {
+//	k, err := NewKey(key)
+//	if err != nil {
+//		panic(fmt.Errorf("could not create label %q: %v", key, err))
+//	}
+//	return Label(k)
+//}
 
 // MustRegister is a helper function that will ensure that the provided Metrics are
 // registered. If a metric fails to register, this method will panic.
@@ -255,97 +241,37 @@ func RegisterIf(metric Metric, enabled func() bool) Metric {
 // NewSum creates a new Metric with an aggregation type of Sum (the values will be cumulative).
 // That means that data collected by the new Metric will be summed before export.
 func NewSum(name, description string, opts ...Options) Metric {
-	return newMetric(name, description, view.Sum(), opts...)
+	return newSum(name, description, opts...)
 }
 
 // NewGauge creates a new Metric with an aggregation type of LastValue. That means that data collected
 // by the new Metric will export only the last recorded value.
 func NewGauge(name, description string, opts ...Options) Metric {
-	return newMetric(name, description, view.LastValue(), opts...)
+	return newGauge(name, description, opts...)
 }
 
 // NewDerivedGauge creates a new Metric with an aggregation type of LastValue that generates the value
 // dynamically according to the provided function. This can be used for values based on querying some
 // state within a system (when event-driven recording is not appropriate).
+//
+// Only 2 usages (uptime and cache expiry in node agent) in istio
 func NewDerivedGauge(name, description string, opts ...DerivedOptions) DerivedMetric {
-	options := createDerivedOptions(opts...)
-	m, err := derivedRegistry.AddFloat64DerivedGauge(name,
-		metric.WithDescription(description),
-		metric.WithLabelKeys(options.labelKeys...),
-		metric.WithUnit(metricdata.UnitDimensionless)) // TODO: allow unit in options
-	if err != nil {
-		log.Warnf("failed to add metric %q: %v", name, err)
-	}
-	derived := &derivedFloat64Metric{
-		base: m,
-		name: name,
-	}
-	if options.valueFn != nil {
-		derived.ValueFrom(options.valueFn)
-	}
-	return derived
+	return newDerivedGauge(name, description, opts...)
 }
 
 // NewDistribution creates a new Metric with an aggregation type of Distribution. This means that the
 // data collected by the Metric will be collected and exported as a histogram, with the specified bounds.
 func NewDistribution(name, description string, bounds []float64, opts ...Options) Metric {
-	return newMetric(name, description, view.Distribution(bounds...), opts...)
+	return newDistribution(name, description, bounds, opts...)
 }
 
-func newMetric(name, description string, aggregation *view.Aggregation, opts ...Options) Metric {
-	o := createOptions(opts...)
-	if o.useInt64 {
-		return newInt64Metric(name, description, aggregation, o)
-	}
-	return newFloat64Metric(name, description, aggregation, o)
-}
-
-type derivedFloat64Metric struct {
-	base *metric.Float64DerivedGauge
-
-	name string
-}
-
-func (d *derivedFloat64Metric) Name() string {
-	return d.name
-}
-
-// no-op
-func (d *derivedFloat64Metric) Register() error {
-	return nil
-}
-
-func (d *derivedFloat64Metric) ValueFrom(valueFn func() float64, labelValues ...string) {
-	if len(labelValues) == 0 {
-		if err := d.base.UpsertEntry(valueFn); err != nil {
-			log.Errorf("failed to add value for derived metric %q: %v", d.name, err)
-		}
-		return
-	}
-	lv := make([]metricdata.LabelValue, 0, len(labelValues))
-	for _, l := range labelValues {
-		lv = append(lv, metricdata.NewLabelValue(l))
-	}
-	if err := d.base.UpsertEntry(valueFn, lv...); err != nil {
-		log.Errorf("failed to add value for derived metric %q: %v", d.name, err)
-	}
-}
-
-type float64Metric struct {
-	*stats.Float64Measure
-
-	// tags stores all tags for the metrics
-	tags []tag.Mutator
-	// ctx is a precomputed context holding tags, as an optimization
-	ctx  context.Context
-	view *view.View
-
-	incrementMeasure []stats.Measurement
-	decrementMeasure []stats.Measurement
-}
+var newSum func(name, description string, opts ...Options) Metric
+var newGauge func(name, description string, opts ...Options) Metric
+var newDistribution func(name, description string, bounds []float64, opts ...Options) Metric
+var newDerivedGauge func(name, description string, opts ...DerivedOptions) DerivedMetric
 
 func createOptions(opts ...Options) *options {
-	o := &options{unit: None, labels: make([]Label, 0)}
+	o := &options{unit: None, labels: make([]string, 0)}
 	for _, opt := range opts {
 		opt(o)
 	}
@@ -363,168 +289,4 @@ func createDerivedOptions(opts ...DerivedOptions) *derivedOptions {
 		o.labelKeys = []string{}
 	}
 	return o
-}
-
-func newFloat64Metric(name, description string, aggregation *view.Aggregation, opts *options) *float64Metric {
-	measure := stats.Float64(name, description, string(opts.unit))
-	tagKeys := make([]tag.Key, 0, len(opts.labels))
-	for _, l := range opts.labels {
-		tagKeys = append(tagKeys, tag.Key(l))
-	}
-	ctx, _ := tag.New(context.Background()) //nolint:errcheck
-	return &float64Metric{
-		Float64Measure:   measure,
-		tags:             make([]tag.Mutator, 0),
-		ctx:              ctx,
-		view:             &view.View{Measure: measure, TagKeys: tagKeys, Aggregation: aggregation},
-		incrementMeasure: []stats.Measurement{measure.M(1)},
-		decrementMeasure: []stats.Measurement{measure.M(-1)},
-	}
-}
-
-func (f *float64Metric) Increment() {
-	f.recordMeasurements(f.incrementMeasure)
-}
-
-func (f *float64Metric) Decrement() {
-	f.recordMeasurements(f.decrementMeasure)
-}
-
-func (f *float64Metric) Name() string {
-	return f.Float64Measure.Name()
-}
-
-func (f *float64Metric) Record(value float64) {
-	recordHookMutex.RLock()
-	if rh, ok := recordHooks[f.Name()]; ok {
-		rh.OnRecordFloat64Measure(f.Float64Measure, f.tags, value)
-	}
-	recordHookMutex.RUnlock()
-	m := f.M(value)
-	stats.Record(f.ctx, m) //nolint:errcheck
-}
-
-func (f *float64Metric) recordMeasurements(m []stats.Measurement) {
-	recordHookMutex.RLock()
-	if rh, ok := recordHooks[f.Name()]; ok {
-		for _, mv := range m {
-			rh.OnRecordFloat64Measure(f.Float64Measure, f.tags, mv.Value())
-		}
-	}
-	recordHookMutex.RUnlock()
-	stats.Record(f.ctx, m...)
-}
-
-func (f *float64Metric) RecordInt(value int64) {
-	f.Record(float64(value))
-}
-
-func (f *float64Metric) With(labelValues ...LabelValue) Metric {
-	t := make([]tag.Mutator, len(f.tags), len(f.tags)+len(labelValues))
-	copy(t, f.tags)
-	for _, tagValue := range labelValues {
-		t = append(t, tag.Mutator(tagValue))
-	}
-	ctx, _ := tag.New(context.Background(), t...) //nolint:errcheck
-	return &float64Metric{
-		Float64Measure:   f.Float64Measure,
-		tags:             t,
-		ctx:              ctx,
-		view:             f.view,
-		incrementMeasure: f.incrementMeasure,
-		decrementMeasure: f.decrementMeasure,
-	}
-}
-
-func (f *float64Metric) Register() error {
-	return view.Register(f.view)
-}
-
-type int64Metric struct {
-	*stats.Int64Measure
-
-	// tags stores all tags for the metrics
-	tags []tag.Mutator
-	// ctx is a precomputed context holding tags, as an optimization
-	ctx  context.Context
-	view *view.View
-
-	// incrementMeasure is a precomputed +1 measurement to avoid extra allocations in Increment()
-	incrementMeasure []stats.Measurement
-	// decrementMeasure is a precomputed -1 measurement to avoid extra allocations in Decrement()
-	decrementMeasure []stats.Measurement
-}
-
-func newInt64Metric(name, description string, aggregation *view.Aggregation, opts *options) *int64Metric {
-	measure := stats.Int64(name, description, string(opts.unit))
-	tagKeys := make([]tag.Key, 0, len(opts.labels))
-	for _, l := range opts.labels {
-		tagKeys = append(tagKeys, tag.Key(l))
-	}
-	ctx, _ := tag.New(context.Background()) //nolint:errcheck
-	return &int64Metric{
-		Int64Measure:     measure,
-		tags:             make([]tag.Mutator, 0),
-		ctx:              ctx,
-		view:             &view.View{Measure: measure, TagKeys: tagKeys, Aggregation: aggregation},
-		incrementMeasure: []stats.Measurement{measure.M(1)},
-		decrementMeasure: []stats.Measurement{measure.M(-1)},
-	}
-}
-
-func (i *int64Metric) Increment() {
-	i.recordMeasurements(i.incrementMeasure)
-}
-
-func (i *int64Metric) Decrement() {
-	i.recordMeasurements(i.decrementMeasure)
-}
-
-func (i *int64Metric) Name() string {
-	return i.Int64Measure.Name()
-}
-
-func (i *int64Metric) Record(value float64) {
-	i.RecordInt(int64(math.Floor(value)))
-}
-
-func (i *int64Metric) recordMeasurements(m []stats.Measurement) {
-	recordHookMutex.RLock()
-	if rh, ok := recordHooks[i.Name()]; ok {
-		for _, mv := range m {
-			rh.OnRecordInt64Measure(i.Int64Measure, i.tags, int64(math.Floor(mv.Value())))
-		}
-	}
-	recordHookMutex.RUnlock()
-	stats.Record(i.ctx, m...) //nolint:errcheck
-}
-
-func (i *int64Metric) RecordInt(value int64) {
-	recordHookMutex.RLock()
-	if rh, ok := recordHooks[i.Name()]; ok {
-		rh.OnRecordInt64Measure(i.Int64Measure, i.tags, value)
-	}
-	recordHookMutex.RUnlock()
-	stats.Record(i.ctx, i.M(value)) //nolint:errcheck
-}
-
-func (i *int64Metric) With(labelValues ...LabelValue) Metric {
-	t := make([]tag.Mutator, len(i.tags), len(i.tags)+len(labelValues))
-	copy(t, i.tags)
-	for _, tagValue := range labelValues {
-		t = append(t, tag.Mutator(tagValue))
-	}
-	ctx, _ := tag.New(context.Background(), t...) //nolint:errcheck
-	return &int64Metric{
-		Int64Measure:     i.Int64Measure,
-		tags:             t,
-		ctx:              ctx,
-		view:             i.view,
-		incrementMeasure: i.incrementMeasure,
-		decrementMeasure: i.decrementMeasure,
-	}
-}
-
-func (i *int64Metric) Register() error {
-	return view.Register(i.view)
 }
